legend option in the draw_option added - now used when drawing the legend - note handling of the original draw option has changed (use striped); added self.pdf() which saves a pdf file of the list canvas debug utils moved to a module (subdir)
